Refactor project name and implement uv for pkg management
