+++ docs/api.md
@@ -1,3 +1,3 @@
 # API Reference
 
-::: llm_toolz.hello_world
+::: llmz.hello_world
