--- src/llm_toolz/nn.py
@@ -1,18 +1,0 @@
-"""Neural network modules."""
-from torch import nn
-
-
-class MultiHeadAttention(nn.Module):
-    pass
-
-
-class GELU(nn.Module):
-    pass
-
-
-class LayerNorm(nn.Module):
-    pass
-
-
-class TransformerBlock(nn.Module):
-    pass
