+++ src/llmz/__init__.py
@@ -1,0 +1,1 @@
+"""The llmz package."""
